Custom Select Box Design
In this we will learn how to make a custom select menu using HTML, CSS and JavaScript step by step.
On my computer screen, you can see this is the select menu design that we are going to create in this video. When I click on this select field, it will open the options and here you can see the icon is rotating.
If I click on any option the same text get selected in the select field.
If I click on Facebook then Facebook is selected here, if I click on YouTube then YouTube get selected here. We will learn to create this custom drop-down select menu using HTML, CSS and JavaScript in this video.

Download Images: https://drive.google.com/file/d/1m91zDrGhCM8gO9oL3cI4psmjbb38ITkf/view